requirements.txt
Pillow>=10.0.0
customtkinter>=5.2.0 